add SepReaderExtensions.IO.Async/Sync.cs
