+++ src/Sep/SepReaderExtensions.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Diagnostics.Contracts;
 using System.IO;
-using System.Threading.Tasks;
 
 namespace nietras.SeparatedValues;
 
@@ -34,143 +33,6 @@ public static partial class SepReaderExtensions
     {
         Contract.Assume(configure != null);
         return configure(Reader(spec.Sep));
-    }
-
-    public static SepReader FromText(this in SepReaderOptions options, string text)
-    {
-        ArgumentNullException.ThrowIfNull(text);
-        var reader = new StringReader(text);
-        Func<SepReader.Info, string> display = static info => $"String Length={((string)info.Source).Length}";
-        return FromWithInfo(new(text, display), options, reader);
-    }
-
-    public static SepReader FromFile(this in SepReaderOptions options, string filePath)
-    {
-        var reader = new StreamReader(filePath, s_streamReaderOptions);
-        Func<SepReader.Info, string> display = static info => $"File='{info.Source}'";
-        return FromWithInfo(new(filePath, display), options, reader);
-    }
-
-    public static SepReader From(this in SepReaderOptions options, byte[] buffer)
-    {
-        var reader = new StreamReader(new MemoryStream(buffer));
-        Func<SepReader.Info, string> display = static info => $"Bytes Length={((byte[])info.Source).Length}";
-        return FromWithInfo(new(buffer, display), options, reader);
-    }
-
-    public static SepReader From(this in SepReaderOptions options, string name, Func<string, Stream> nameToStream)
-    {
-        ArgumentNullException.ThrowIfNull(nameToStream);
-        var reader = new StreamReader(nameToStream(name));
-        Func<SepReader.Info, string> display = static info => $"Stream Name='{info.Source}'";
-        return FromWithInfo(new(name, display), options, reader);
-    }
-    public static SepReader From(this in SepReaderOptions options, Stream stream)
-    {
-        var reader = new StreamReader(stream);
-        Func<SepReader.Info, string> display = static info => $"Stream='{info.Source}'";
-        return FromWithInfo(new(stream, display), options, reader);
-    }
-
-    public static SepReader From(this in SepReaderOptions options, string name, Func<string, TextReader> nameToReader)
-    {
-        ArgumentNullException.ThrowIfNull(nameToReader);
-        var reader = nameToReader(name);
-        Func<SepReader.Info, string> display = static info => $"TextReader Name='{info.Source}'";
-        return FromWithInfo(new(name, display), options, reader);
-    }
-    public static SepReader From(this in SepReaderOptions options, TextReader reader)
-    {
-        Func<SepReader.Info, string> display = static info => $"TextReader='{info.Source}'";
-        return FromWithInfo(new(reader, display), options, reader);
-    }
-
-    internal static SepReader FromWithInfo(SepReader.Info info, SepReaderOptions options, TextReader reader)
-    {
-        ArgumentNullException.ThrowIfNull(reader);
-        SepReader? sepReader = null;
-        try
-        {
-            sepReader = new SepReader(info, options, reader);
-            sepReader.Initialize(options);
-            return sepReader;
-        }
-        catch (Exception)
-        {
-            sepReader?.Dispose();
-            reader.Dispose();
-            throw;
-        }
-    }
-
-    public static async ValueTask<SepReader> FromTextAsync(this SepReaderOptions options, string text)
-    {
-        ArgumentNullException.ThrowIfNull(text);
-        var reader = new StringReader(text);
-        Func<SepReader.Info, string> display = static info => $"String Length={((string)info.Source).Length}";
-        return await FromWithInfoAsync(new(text, display), options, reader);
-    }
-
-    public static async ValueTask<SepReader> FromFileAsync(this SepReaderOptions options, string filePath)
-    {
-        var reader = new StreamReader(filePath, s_streamReaderOptions);
-        Func<SepReader.Info, string> display = static info => $"File='{info.Source}'";
-        return await FromWithInfoAsync(new(filePath, display), options, reader);
-    }
-
-    public static async ValueTask<SepReader> FromAsync(this SepReaderOptions options, byte[] buffer)
-    {
-        var reader = new StreamReader(new MemoryStream(buffer));
-        Func<SepReader.Info, string> display = static info => $"Bytes Length={((byte[])info.Source).Length}";
-        return await FromWithInfoAsync(new(buffer, display), options, reader);
-    }
-
-    public static async ValueTask<SepReader> FromAsync(this SepReaderOptions options, string name, Func<string, Stream> nameToStream)
-    {
-        ArgumentNullException.ThrowIfNull(nameToStream);
-        var reader = new StreamReader(nameToStream(name));
-        Func<SepReader.Info, string> display = static info => $"Stream Name='{info.Source}'";
-        return await FromWithInfoAsync(new(name, display), options, reader);
-    }
-
-    public static async ValueTask<SepReader> FromAsync(this SepReaderOptions options, Stream stream)
-    {
-        var reader = new StreamReader(stream);
-        Func<SepReader.Info, string> display = static info => $"Stream='{info.Source}'";
-        return await FromWithInfoAsync(new(stream, display), options, reader);
-    }
-
-    public static async ValueTask<SepReader> FromAsync(this SepReaderOptions options, string name, Func<string, TextReader> nameToReader)
-    {
-        ArgumentNullException.ThrowIfNull(nameToReader);
-        var reader = nameToReader(name);
-        Func<SepReader.Info, string> display = static info => $"TextReader Name='{info.Source}'";
-        return await FromWithInfoAsync(new(name, display), options, reader);
-    }
-
-    public static async ValueTask<SepReader> FromAsync(this SepReaderOptions options, TextReader reader)
-    {
-        Func<SepReader.Info, string> display = static info => $"TextReader='{info.Source}'";
-        return await FromWithInfoAsync(new(reader, display), options, reader);
-    }
-
-    internal static async ValueTask<SepReader> FromWithInfoAsync(SepReader.Info info, SepReaderOptions options, TextReader reader)
-    {
-        ArgumentNullException.ThrowIfNull(reader);
-        SepReader? sepReader = null;
-        try
-        {
-            sepReader = new SepReader(info, options, reader);
-            await sepReader.InitializeAsync(options, default);
-            return sepReader;
-        }
-        catch (Exception)
-        {
-            // There is no DisposeAsync on TextReader so none on SepReader either
-            sepReader?.Dispose();
-            reader.Dispose();
-            throw;
-        }
     }
 
     internal static Sep? DetectSep(ReadOnlySpan<char> chars, bool nothingLeftToRead, bool disableQuotesParsing)
